refactor: moves OAuth client creation function into lib
