+++ src/lib/oauth2.ts
@@ -1,3 +1,5 @@
+import { AuthorizationCode } from 'simple-oauth2';
+
 export const config = {
   client: {
     id: process.env.OAUTH_CLIENT_ID,
@@ -9,3 +11,27 @@ export const config = {
     authorizePath: `/login/oauth/authorize`,
   },
 };
+
+/**
+ * Creates a new client for OAuth2 Authorization Code grant type.
+ * See https://github.com/lelylan/simple-oauth2/blob/master/API.md#new-authorizationcodeoptions
+ */
+export const createClient = () => {
+  return new AuthorizationCode(config);
+};
+
+export const renderBody = (status: string, content: Object) => `
+  <script>
+    const receiveMessage = (message) => {
+      window.opener.postMessage(
+        'authorization:github:${status}:${JSON.stringify(content)}',
+        message.origin
+      );
+  
+      window.removeEventListener("message", receiveMessage, false);
+    }
+    window.addEventListener("message", receiveMessage, false);
+    
+    window.opener.postMessage("authorizing:github", "*");
+  </script>
+  `;
