+++ src/pages/api/callback.ts
@@ -1,22 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { config } from '@/lib/oauth2';
-import { AuthorizationCode } from 'simple-oauth2';
-
-const renderBody = (status: string, content: Object) => `
-  <script>
-    const receiveMessage = (message) => {
-      window.opener.postMessage(
-        'authorization:github:${status}:${JSON.stringify(content)}',
-        message.origin
-      );
-  
-      window.removeEventListener("message", receiveMessage, false);
-    }
-    window.addEventListener("message", receiveMessage, false);
-    
-    window.opener.postMessage("authorizing:github", "*");
-  </script>
-  `;
+import { renderBody, createClient } from '@/lib/oauth2';
 
 export default async function handler(
   req: NextApiRequest,
@@ -25,28 +8,17 @@ export default async function handler(
   const code = req.query.code;
   const { host } = req.headers;
 
-  const oauthClient = new AuthorizationCode({
-    client: {
-      id: process.env.OAUTH_CLIENT_ID,
-      secret: process.env.OAUTH_CLIENT_SECRET,
-    },
-    auth: {
-      tokenHost: `https://github.com`,
-      tokenPath: `/login/oauth/access_token`,
-      authorizePath: `/login/oauth/authorize`,
-    },
-  });
+  const oauthClient = createClient();
 
   try {
-    const accessToken = await oauthClient.authorizationCode.getToken({
+    const accessToken = await oauthClient.getToken({
       code,
       redirect_uri: `https://${host}/api/callback`,
     });
-    const { token } = oauthClient.accessToken.create(accessToken);
 
     res.status(200).send(
       renderBody('success', {
-        token: token.access_token,
+        token: accessToken,
         provider: 'github',
       })
     );
